Added LR parser engine
What on earth posessed me to reimplement Flex and Bison in
Javascript, but worse? I do not have a good excuse. NIH syndrome?
Maybe, but that describes most of this project.

Maybe I just saw this an an opportunity to finally learn how
parsers worked. In that regard, I'd say it was a success.

I really need to clean up the code, though. I also need to add
unit tests.